# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumen — Catholic Prayer Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wall:
    get:
      operationId: listWallPosts
      tags: [wall]
      summary: List approved Wall of Hope posts
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [healing, family, disaster, work, vocation, gratitude, other]
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallPostList"
    post:
      operationId: createWallPost
      tags: [wall]
      summary: Submit a new prayer message to the Wall of Hope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWallPostBody"
      responses:
        "201":
          description: Created (pending moderation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallPost"
        "429":
          description: Too many submissions
  /wall/{id}/pray:
    post:
      operationId: prayForWallPost
      tags: [wall]
      summary: Increment "I'm praying for this" counter
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WallPost"
...
components:
...
  schemas:
...
    WallPost:
      type: object
      required: [id, displayName, message, category, prayCount, status, createdAt]
      properties:
        id: { type: integer }
        displayName: { type: string, description: "Anonymous => 'Anonymous'" }
        message: { type: string }
        category:
          type: string
          enum: [healing, family, disaster, work, vocation, gratitude, other]
        prayCount: { type: integer }
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt: { type: string, format: date-time }
    WallPostList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WallPost"
        total: { type: integer }
    CreateWallPostBody:
      type: object
      required: [message, category, anonymous]
      properties:
        displayName: { type: string, nullable: true, maxLength: 60 }
        message: { type: string, minLength: 8, maxLength: 600 }
        category:
          type: string
          enum: [healing, family, disaster, work, vocation, gratitude, other]
        anonymous: { type: boolean }
...